# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Arokah API
  version: 1.0.0
  description: AI-powered personal stylist — daily looks, trip planning, cultural
    intelligence.
paths:
...
  /api/calendar/apple/disconnect/:
    post:
      operationId: calendar_apple_disconnect_create
      description: Removes stored Apple credentials from Arokah.
      summary: Disconnect Apple Calendar
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/calendar/apple/sync/:
    post:
      operationId: calendar_apple_sync_create
      description: Pulls events via CalDAV from the last 7 days and next 60 days.
      summary: Sync Apple Calendar events
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppleSyncResult'
          description: ''
...
  /api/calendar/google/disconnect/:
    post:
      operationId: calendar_google_disconnect_create
      description: Revokes OAuth tokens and removes the connection.
      summary: Disconnect Google Calendar
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/calendar/outlook/disconnect/:
    post:
      operationId: calendar_outlook_disconnect_create
      description: POST /api/calendar/outlook/disconnect/
      summary: Disconnect Outlook Calendar
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/calendar/outlook/sync/:
    post:
      operationId: calendar_outlook_sync_create
      description: Pulls events via Microsoft Graph API for the last 7 days and next
        60 days.
      summary: Sync Outlook Calendar events
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutlookSyncResult'
          description: ''
  /api/calendar/status/:
    get:
      operationId: calendar_status_retrieve
      description: Returns connection status for Google and Apple calendars.
      summary: Calendar connection status
      tags:
      - calendar
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    AppleSyncResult:
      type: object
      properties:
        created:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
      - created
      - errors
      - updated
...
    OutlookSyncResult:
      type: object
      properties:
        created:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
      - created
      - errors
      - updated
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT